backend/requirements.txt
Django==3.1
django-cors-headers==3.5.0
djangorestframework==3.11.1
gunicorn==20.0.4
psycopg2-binary==2.8.5
pytz==2021.3
sqlparse==0.4.2
joblib==0.16.0
gunicorn==20.0.4
Pillow==8.3.1
